# contracts/openapi/link-service.yaml
openapi: 3.1.0
info:
  title: getlink_dtd Profile and Link API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /profiles/{username}:
    get:
      operationId: getPublicProfile
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string, pattern: '^[a-z0-9-]{3,30}$' }
      responses:
        '200':
          description: Public profile and enabled links
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { $ref: '#/components/responses/Problem' }
  /me/profile:
    get:
      operationId: getOwnedProfile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Owned profile; created from signed JWT claims on first access
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateOwnedProfile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /me/links:
    get:
      operationId: listOwnedLinks
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All owned links
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Link' }
        '401': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createLink
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLinkRequest' }
      responses:
        '201':
          description: Link created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Link' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /me/links/{linkId}:
    patch:
      operationId: updateLink
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LinkId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLinkRequest' }
      responses:
        '200':
          description: Link updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Link' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteLink
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LinkId' }]
      responses:
        '204': { description: Link deleted }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /me/links/order:
    put:
      operationId: reorderLinks
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReorderLinksRequest' }
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Link' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /r/{linkId}:
    get:
      operationId: redirectToSavedLink
      parameters: [{ $ref: '#/components/parameters/LinkId' }]
      responses:
        '302': { description: Redirect to validated stored destination }
        '404': { $ref: '#/components/responses/Problem' }
  /me/qr/profile:
    get:
      operationId: getProfileQrCode
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: PNG or SVG profile QR code }
  /me/links/{linkId}/qr:
    get:
      operationId: getLinkQrCode
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LinkId' }]
      responses:
        '200': { description: PNG or SVG link QR code }
  /me/analytics/summary:
    get:
      operationId: getAnalyticsSummary
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Profile views, clicks, and CTR }
  /me/analytics/timeseries:
    get:
      operationId: getThirtyDayAnalytics
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Aggregated daily views and clicks }
  /admin/profiles/{userId}:
    get:
      operationId: getProfileForAdministration
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Profile and links for administration }
        '403': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    LinkId:
      in: path
      name: linkId
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Profile:
      type: object
      required: [id, username, displayName, status, version, links]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        bio: { type: [string, 'null'] }
        avatarUrl: { type: [string, 'null'], format: uri }
        status: { type: string }
        version: { type: integer, format: int64 }
        links:
          type: array
          items: { $ref: '#/components/schemas/Link' }
    Link:
      type: object
      required: [id, title, destinationUrl, position, enabled, clickCount, version]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        destinationUrl: { type: string, format: uri }
        icon: { type: [string, 'null'] }
        position: { type: integer, minimum: 0 }
        enabled: { type: boolean }
        clickCount: { type: integer, format: int64, minimum: 0 }
        version: { type: integer, format: int64 }
    UpdateProfileRequest:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        bio: { type: [string, 'null'], maxLength: 500 }
    CreateLinkRequest:
      type: object
      required: [title, destinationUrl]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        destinationUrl: { type: string, format: uri, maxLength: 2048 }
        icon: { type: [string, 'null'], maxLength: 100 }
    UpdateLinkRequest:
      type: object
      required: [title, destinationUrl, enabled]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        destinationUrl: { type: string, format: uri, maxLength: 2048 }
        icon: { type: [string, 'null'], maxLength: 100 }
        enabled: { type: boolean }
    ReorderLinksRequest:
      type: object
      required: [linkIds]
      properties:
        linkIds:
          type: array
          minItems: 1
          maxItems: 30
          uniqueItems: true
          items: { type: string, format: uuid }
    Problem:
      type: object
      required: [title, status, code, requestId]
      properties:
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }